Move export::{Instance,RefInstance} -> object
