+++ gdnative-async/src/method.rs
@@ -5,10 +5,10 @@ use std::sync::Arc;
 use futures_task::{LocalFutureObj, LocalSpawn, SpawnError};
 
 use gdnative_core::core_types::{ToVariant, Variant};
-use gdnative_core::export::{Method, Varargs};
-use gdnative_core::export::{NativeClass, RefInstance};
+use gdnative_core::export::{Method, NativeClass, Varargs};
 use gdnative_core::log::{self, Site};
 use gdnative_core::object::ownership::Shared;
+use gdnative_core::object::RefInstance;
 
 use crate::rt::Context;
