+++ gdnative-async/src/rt/func_state.rs
@@ -2,11 +2,12 @@ use gdnative_bindings::Reference;
 use gdnative_core::core_types::{ToVariant, Variant, VariantType};
 use gdnative_core::export::user_data::{LocalCellData, Map, MapMut};
 use gdnative_core::export::{
-    ClassBuilder, ExportInfo, PropertyUsage, Signal, SignalArgument, StaticArgs, StaticArgsMethod,
+    ClassBuilder, ExportInfo, NativeClass, NativeClassMethods, PropertyUsage, Signal,
+    SignalArgument, StaticArgs, StaticArgsMethod,
 };
-use gdnative_core::export::{Instance, NativeClass, NativeClassMethods, RefInstance};
 use gdnative_core::godot_site;
 use gdnative_core::object::ownership::{Shared, Unique};
+use gdnative_core::object::{Instance, RefInstance};
 use gdnative_derive::FromVarargs;
 
 use crate::future::Resume;
