+++ gdnative-bindings/src/utils.rs
@@ -1,9 +1,10 @@
 //! Utility functions and extension traits that depend on generated bindings
 
-use super::generated::{Engine, Node, SceneTree};
-use gdnative_core::export::{NativeClass, RefInstance};
+use gdnative_core::export::NativeClass;
 use gdnative_core::object::ownership::Shared;
-use gdnative_core::object::{SubClass, TRef};
+use gdnative_core::object::{RefInstance, SubClass, TRef};
+
+use super::generated::{Engine, Node, SceneTree};
 
 /// Convenience method  to obtain a reference to an "auto-load" node, that is a child of the root
 /// node. Returns `None` if the node does not exist or is not of the correct type.
